Always require a password for simple auth
When building a connection string with simple auth
we now always require a password. The is necessary because
ARM won't provide a password for vCore clusters unlike
it does for mongo accounts, since the password is
set by the user and encrypted.

In addition if ADS has once successfully connected to the
resource it will preserve the password through IConnectionOptions
so there is no need to re-query it from the keychain.